requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.6
Markdown==2.6.11
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.4
requests==2.18.4
urllib3==1.22
